# FaxtimeRest/config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "title": "TAIHOINST WEB API",
        "description": "A TAIHOINST WEB API based on Swagger and built using Node.js"
    },
    "paths": {
...
        "/v1/others/file/upload" : {
            "post" : {
                "tags": [
                    "/V1/OTHERS/FILE_UPLOAD"
                ],
                "operationId": "OTHERS_FILE_UPLOAD_POST",
                "consumes": [
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json",
                    "text/html"
                ],
                "parameters": [
                    { "$ref" : "#/parameters/agentParam" },
                    {
                        "name": "objName",
                        "in": "query",
                        "description": "name...",
                        "required": true,
                        "type": "string",
                        "default" : "mmsfile1"
                    },
                    {
                        "name": "maxByte",
                        "in": "query",
                        "description": "byte...",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "subPath",
                        "in": "query",
                        "description": "path...",
                        "required": false,
                        "type": "string",
                        "default" : "mmsfile"
                    },
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "첨부파일",
                        "required": true,
                        "type": "file"
                    }
                ],
                "description": "파일업로드",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/RESULTS"
                            }
                        }
                    }
                }
            }
        }
    },
    "parameters": {
        "agentParam" : {
            "name": "Agent-Key",
            "in": "header",
            "description": "agent Identifier...",
            "required": true,
            "type": "string",
            "default" : "30748F97-429A-443B-BC66-9A32AD20B07F"
        },
...
    },
    "securityDefinitions": {
        "apikeyQuery" : {
            "type": "apiKey",
            "in": "header",
            "name": "QUERY-API-KEY"
        },
        "userSecurity": {
            "type": "apiKey",
            "in": "header",
            "name": "USER-API-KEY"
        },
        "adminSecurity": {
            "type": "apiKey",
            "in": "header",
            "name": "ADMIN-API-KEY"
        },
        "agentSecurity": {
            "type": "apiKey",
            "in": "header",
            "name": "AGENT-API-KEY"
        },
        "mediaSecurity": {
            "type": "apiKey",
            "in": "query",
            "name": "MEDIA-API-KEY"
        }
    },
    "definitions": {
        "RESULTS": {
            "type": "object",
            "properties": {
                "status" : {
                    "type" : "string"
                },
                "description" : {
                    "type" : "string"
                },
                "datas" : {
                    "type" : "object"
                }
            }
        },
...
    }
}